chains base service refactor; WIP; #147
